+++ CHANGELOG.md
@@ -9,6 +9,7 @@ BREAKING CHANGES
 - [core] Add Coin's MaxSupply
 - [core] Remove CreatedAtBlock field in candidates
 - [core] Add GasCoin to Checks
+- [core] Fix buy coin commission calculation
 - [gui] Remove GUI
 - [config] KeepStateHistory -> KeepLastStates
 - [config] Add state_cache_size option